Fixed some warnings.

+++ xpilot/src/client/protoclient.h
@@ -95,7 +95,7 @@ extern int Init_asteroids(void);
 /*
  * query.c
  */
-extern int Query_all(sock_t *sockfd, int port, char *msg, int msglen);
+extern int Query_all(sock_t *sockfd, int port, char *msg, size_t msglen);
